change canBeReturndForDays to int

+++ src/main/kotlin/org/misarch/catalog/event/model/ProductVariantVersionDTO.kt
@@ -11,7 +11,7 @@ import java.util.UUID
  * @property version version of the product variant version
  * @property retailPrice retail price of the product variant version
  * @property createdAt creation timestamp of the product variant version
- * @property canBeReturnedForDays number of days the product variant version can be returned, null if it cannot be returned
+ * @property canBeReturnedForDays number of days the product variant version can be returned, null if it can be returned indefinitely
  * @property productVariantId id of the referenced product variant
  * @property taxRateId id of the referenced tax rate
  * @property weight weight of the product variant version
@@ -23,7 +23,7 @@ data class ProductVariantVersionDTO(
     val version: Int,
     val retailPrice: Int,
     val createdAt: String,
-    val canBeReturnedForDays: Double?,
+    val canBeReturnedForDays: Int?,
     val productVariantId: UUID,
     val taxRateId: UUID,
     val weight: Double
